requirements.txt
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
build==1.2.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-livetw==0.4.2
Flask-Login==0.6.3
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.1
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyproject_hooks==1.1.0
pytailwindcss==0.2.0
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.19.1
SQLAlchemy==2.0.31
sqlalchemy-cockroachdb==2.0.2
tomli==2.0.1
typing_extensions==4.12.2
validator-collection==1.5.0
websockets==12.0
Werkzeug==3.0.3
